fix(artifacts): minor refactoring
